third_party/WebKit/LayoutTests/fast/dom/shadow/content-deep-expected.txt

CONSOLE WARNING: line 1: /deep/ combinator is deprecated and will be a no-op in M60, around August 2017. See https://www.chromestatus.com/features/4964279606312960 for more details.
::content /deep/ should match deep into distributed subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(target).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
